+++ Changes.md
@@ -5,6 +5,20 @@ Improvements
 ------------
 
 - Arnold : Disabled ADP more thoroughly, using `ARNOLD_ADP_DISABLE=1` for Arnold 7.1.4.0 and greater. Set `ARNOLD_ADP_DISABLE=0` before running Gaffer if you wish to use ADP.
+- Cycles :
+  - Added warning when a shader contains a parameter which does not exist in Cycles.
+  - Added warning when a shader contains a parameter with an unsupported datatype.
+  - Added warning when a shader contains an enum parameter with an invalid value.
+  - Added support for passing InternedStringData to enum and string parameters.
+
+Fixes
+-----
+
+- Cyles :
+  - Fixed crashes caused by providing unsupported data types in shader parameters.
+  - Fixed support for Color4f values on colour shader parameters. This can be useful when loading non-standard USD files.
+  - Fixed support for V[23]i values on vector shader parameters.
+  - Fixed handling of colour array parameters.
 
 1.1.8.0 (relative to 1.1.7.0)
 =======